+++ R/densitygamma.R
@@ -1,20 +1,18 @@
-densitygamma <- function(x, smoothness, from, to, cut) {
-	pi <- x$pi
-	mu <- x$mu
-	sd <- x$sd
-	alpha <- x$alpha
-	lambda <- x$lambda
-	if(missing(from)) {
-		from <- max(min(mu - cut * sd), 0)
-	}
-	if(missing(to)) {
-		to <- max(mu + 1.2 * cut * sd)
-	}
-	xseq <- seq(from, to, length = smoothness)
-    res <- matrix(NA, nrow = length(xseq), ncol = length(pi))
-	for(i in 1:length(pi)) {
-		res[ ,i] <-  pi[i] * dgamma(xseq, alpha[i], lambda[i])
-	}
-	yt <- apply(res, 1, sum)
-    structure(list(x = xseq, y = yt), class = "densityEM")
+densitygamma <- function(x, at, smoothness, cut) {
+  pi <- x$pi
+  mu <- x$mu
+  sd <- x$sd
+  alpha <- x$alpha
+  lambda <- x$lambda
+  if(missing(at)) {
+    from <- max(min(mu - cut * sd), 0)
+    to <- max(mu + 1.1 * cut * sd)
+    at <- seq(from, to, length = smoothness)
+  }
+  res <- matrix(NA, nrow = length(at), ncol = length(pi))
+  for(i in 1:length(pi)) {
+    res[ ,i] <-  pi[i] * dgamma(at, alpha[i], lambda[i])
+  }
+  y <- apply(res, 1, sum)
+  structure(list(x = at, y = y, comp = res), class = "densityEM")
 }