+++ R/densitylnorm.R
@@ -1,24 +1,18 @@
-densitylnorm <- function(x, smoothness, from, to, cut) {
-	pi <- x$pi
-	mu <- x$mu
-	sd <- x$sd
-	mulog <- x$mulog
-	sdlog <- x$sdlog
-	if(missing(from)) {
-		from <- max(min(mu - cut * sd), 0)
-	}
-	if(missing(to)) {
-		to <- max(mu + 1.2 * cut * sd)
-	}
-	xseq <- seq(from, to, length = smoothness)
-    res <- matrix(NA, nrow = length(xseq), ncol = length(pi))
-	for(i in 1:length(pi)) {
-		res[ ,i] <-  pi[i] * dlnorm(xseq, mulog[i], sdlog[i])
-	}
-	yt <- apply(res, 1, sum)
-    structure(list(x = xseq, y = yt), class = "densityEM")
+densitylnorm <- function(x, at, smoothness, cut) {
+  pi <- x$pi
+  mu <- x$mu
+  sd <- x$sd
+  mulog <- x$mulog
+  sdlog <- x$sdlog
+  if(missing(at)) {
+    from <- max(min(mu - cut * sd), 0)
+    to <- max(mu + 1.1 * cut * sd)
+    at <- seq(from, to, length = smoothness)
+  }
+  res <- matrix(NA, nrow = length(at), ncol = length(pi))
+  for(i in 1:length(pi)) {
+    res[ ,i] <-  pi[i] * dlnorm(at, mulog[i], sdlog[i])
+  }
+  y <- apply(res, 1, sum)
+  structure(list(x = at, y = y, comp = res), class = "densityEM")
 }
-
-
-
-