+++ NEWS.md
@@ -1,4 +1,10 @@
 
+## 0.2.0
+
+- Updated the plot and density function
+- Solved the NA issue in the output of mixfit() (NaN still happens when EM algorithm fails to converge)
+- Added a vignette
+
 ## 0.1.1
 
 - Resolved the warnings issues found in [check results](https://www.r-project.org/nosvn/R.check/r-release-windows-ix86+x86_64/mixR-00check.html).